moved submodule class out of __init__.py
